04_Dashboard/requirements_dashboard.txt
streamlit==1.28.1
pandas==2.2.0
plotly==5.18.0
psycopg2-binary==2.9.9
numpy==1.24.3
sqlalchemy==2.0.25
